+++ src/main/java/com/telepathicgrunt/blame/main/SpawnHelperBlame.java
@@ -24,12 +24,12 @@ import org.apache.logging.log4j.Level;
  */
 public class SpawnHelperBlame {
 
-    public static void addMobCrashDetails(ServerWorld serverWorld, SpawnGroup entityClassification, BlockPos pos, Biome biome, Pool<SpawnSettings.SpawnEntry> pool) {
+    public static void addMobCrashDetails(ServerWorldAccess serverWorld, SpawnGroup entityClassification, BlockPos pos, Biome biome, Pool<SpawnSettings.SpawnEntry> pool) {
         // Figure out if mob spawning is gonna crash game and to run our code if so.
         int totalWeight = ((PoolAccessor)pool).blame_getTotalWeight();
         if (totalWeight < 0) {
 
-            RegistryKey<World> worldID = serverWorld.getRegistryKey();
+            RegistryKey<World> worldID = serverWorld.toServerWorld().getRegistryKey();
             Identifier biomeID = serverWorld.getRegistryManager().get(Registry.BIOME_KEY).getId(biome);
 
             // Add extra info to the log file.
@@ -46,23 +46,22 @@ public class SpawnHelperBlame {
         }
     }
 
-    public static void addMobCrashDetails(ServerWorldAccess serverWorld, ChunkPos chunkPos, Biome biome, Pool<SpawnSettings.SpawnEntry> pool) {
+    public static void addMobGroupCrashDetails(ServerWorldAccess serverWorld, SpawnGroup entityClassification, BlockPos pos, Biome biome, SpawnSettings.SpawnEntry spawner) {
         // Figure out if mob spawning is gonna crash game and to run our code if so.
-        int totalWeight = ((PoolAccessor)pool).blame_getTotalWeight();
-        if (totalWeight < 0) {
+        if (spawner != null && spawner.minGroupSize > spawner.maxGroupSize) {
 
             RegistryKey<World> worldID = serverWorld.toServerWorld().getRegistryKey();
             Identifier biomeID = serverWorld.getRegistryManager().get(Registry.BIOME_KEY).getId(biome);
 
             // Add extra info to the log file.
-            Blame.LOGGER.log(Level.ERROR, "\n****************** Blame Report Mob Weights " + Blame.VERSION + " ******************" +
-                    "\n  Detected total weight of mob list is negative which will crash the game! " +
+            Blame.LOGGER.log(Level.ERROR, "\n\n****************** Blame Report Mob Groups " + Blame.VERSION + " ******************" +
+                    "\n  Detected a mob with a minGroup size larger than maxGroup size which will crash the game! minGroup must always be smaller." +
                     "\n  See info below to find which mob is the problem and where it is attempting to spawn at." +
-                    "\n World Registry Name : " + worldID.getValue().toString() +
+                    "\n World Registry Name : " + worldID.getValue() +
                     "\n Biome Registry Name : " + (biomeID != null ? biomeID.toString() : "Wait what? How is the biome not registered and has no registry name!?!? This should be impossible!!!") +
-                    "\n Classification of entity being spawned : " + SpawnGroup.CREATURE.getName() +
-                    "\n Entity position :  chunk " + chunkPos.x + " " + chunkPos.z +
-                    "\n Weighted list of mobs to spawn : " + printMobListContents(pool) +
+                    "\n Classification of entity being spawned : " + entityClassification.getName() +
+                    "\n Entity position : " + pos.toString() +
+                    "\n Mobs attempted to be spawn : " + spawner +
                     "\n");
 
         }