+++ src/main/java/com/telepathicgrunt/blame/mixin/SinglePoolElementAccessor.java
@@ -3,9 +3,12 @@ package com.telepathicgrunt.blame.mixin;
 import com.mojang.datafixers.util.Either;
 import net.minecraft.structure.Structure;
 import net.minecraft.structure.pool.SinglePoolElement;
+import net.minecraft.structure.processor.StructureProcessorList;
 import net.minecraft.util.Identifier;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.function.Supplier;
 
 /* @author - TelepathicGrunt
  *
@@ -18,4 +21,7 @@ public interface SinglePoolElementAccessor {
 
     @Accessor("location")
     Either<Identifier, Structure> blame_getTemplateID();
+
+    @Accessor("processors")
+    Supplier<StructureProcessorList> blame_getProcessors();
 }