+++ src/main/java/com/telepathicgrunt/blame/mixin/SpawnHelperMixin.java
@@ -2,6 +2,7 @@ package com.telepathicgrunt.blame.mixin;
 
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.structure.StructureManager;
 import net.minecraft.util.collection.Pool;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
@@ -19,10 +20,12 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.Random;
 
 import static com.telepathicgrunt.blame.main.SpawnHelperBlame.addMobCrashDetails;
+import static com.telepathicgrunt.blame.main.SpawnHelperBlame.addMobGroupCrashDetails;
 
 /* @author - TelepathicGrunt
  *
@@ -36,7 +39,7 @@ public class SpawnHelperMixin {
     @Redirect(method = "pickRandomSpawnEntry(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Lnet/minecraft/entity/SpawnGroup;Ljava/util/Random;Lnet/minecraft/util/math/BlockPos;)Ljava/util/Optional;",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/util/collection/Pool;getOrEmpty(Ljava/util/Random;)Ljava/util/Optional;"))
     private static Optional<SpawnSettings.SpawnEntry> blame_checkIfMobSpawnWillCrash1(Pool<SpawnSettings.SpawnEntry> pool, Random random,
-                                                                                     ServerWorld serverWorld, StructureAccessor structureAccessor,
+                                                                                      ServerWorld serverWorld, StructureAccessor structureAccessor,
                                                                                      ChunkGenerator chunkGenerator, SpawnGroup spawnGroup,
                                                                                      Random random2, BlockPos pos) {
         Biome biome = serverWorld.getBiome(pos);
@@ -48,6 +51,29 @@ public class SpawnHelperMixin {
             at = @At(value = "INVOKE", target = "Lnet/minecraft/util/collection/Pool;getOrEmpty(Ljava/util/Random;)Ljava/util/Optional;", ordinal = 0),
             locals = LocalCapture.CAPTURE_FAILHARD)
     private static void blame_checkIfMobSpawnWillCrash2(ServerWorldAccess world, Biome biome, ChunkPos chunkPos, Random random, CallbackInfo ci, SpawnSettings spawnSettings, Pool<SpawnSettings.SpawnEntry> pool) {
-        addMobCrashDetails(world, chunkPos, biome, pool);
+        addMobCrashDetails(world, SpawnGroup.CREATURE, chunkPos.getStartPos(), biome, pool);
+    }
+
+    @Inject(method = "pickRandomSpawnEntry(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Lnet/minecraft/entity/SpawnGroup;Ljava/util/Random;Lnet/minecraft/util/math/BlockPos;)Ljava/util/Optional;",
+            at = @At(value = "RETURN"),
+            locals = LocalCapture.CAPTURE_FAILHARD)
+    private static void blame_checkIfMobSpawnWillCrash3(ServerWorld world, StructureAccessor structureAccessor,
+                                                        ChunkGenerator chunkGenerator, SpawnGroup spawnGroup,
+                                                        Random random, BlockPos pos,
+                                                        CallbackInfoReturnable<Optional<SpawnSettings.SpawnEntry>> cir,
+                                                        Biome biome) {
+        if(cir.getReturnValue().isPresent()) {
+            addMobGroupCrashDetails(world, spawnGroup, pos, biome, cir.getReturnValue().get());
+        }
+    }
+
+    @Inject(method = "populateEntities(Lnet/minecraft/world/ServerWorldAccess;Lnet/minecraft/world/biome/Biome;Lnet/minecraft/util/math/ChunkPos;Ljava/util/Random;)V",
+            at = @At(value = "INVOKE", ordinal = 0, remap = false, target = "Ljava/util/Random;nextInt(I)I"),
+            locals = LocalCapture.CAPTURE_FAILHARD)
+    private static void blame_checkIfMobSpawnWillCrash4(ServerWorldAccess world, Biome biome, ChunkPos chunkPos,
+                                                        Random random, CallbackInfo ci, SpawnSettings mobspawninfo,
+                                                        Pool<SpawnSettings.SpawnEntry> pool, int chunkX, int chunkZ,
+                                                        Optional<SpawnSettings.SpawnEntry> optional, SpawnSettings.SpawnEntry spawner) {
+        addMobGroupCrashDetails(world, SpawnGroup.CREATURE, new ChunkPos(chunkX, chunkZ).getStartPos(), biome, spawner);
     }
 }